#296 #306 use ARRAY type multi-valued cols. some docu, option -gt 65535

+++ bagv2/etl/sql/show-db-stats.sql
@@ -19,7 +19,7 @@ ORDER BY schema;
 
 -- Row counts per schema en per tabel
 SELECT
-  nspname AS schemaname,relname,reltuples
+  nspname AS schemaname,relname,reltuples::bigint
 FROM pg_class C
 LEFT JOIN pg_namespace N ON (N.oid = C.relnamespace)
 WHERE
